move tests into functional groupings
