+++ routes/graphql/schema.graphql
@@ -57,6 +57,8 @@ type Project {
     repository: String!
     created_at: DateTime!
     updated_at: DateTime
+
+    latestBuild: Build
 }
 
 type Team {
