--- tests/Feature/QueriesTest.php
@@ -1,396 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-use App\Models\Build;
-use App\Models\Project;
-use App\Models\Team;
-use App\Models\User;
-use DB;
-use Tests\PassportTestCase;
-use Illuminate\Foundation\Testing\WithFaker;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Carbon;
-
-class QueriesTest extends PassportTestCase
-{
-    public function testBuildsQuery()
-    {
-        $builds = factory(Build::class, 12)
-            ->create()
-            ->take(10)
-            ->map(
-                function ($build) {
-                    return [
-                        'id' => (string) $build->id,
-                        'hash' => $build->hash,
-                        'project' => [
-                            'id' => (string) $build->project->id,
-                            'name' => $build->project->name,
-                            'repository' => $build->project->repository,
-                        ],
-                        'status' => $build->status,
-                        'commit' => $build->commit,
-                        'completed_at' => $build->completed_at,
-                    ];
-                }
-            )
-            ->all();
-
-        $expected = [
-            'data' => [
-                'builds' => [
-                    'data' => $builds,
-                    'paginatorInfo' => [
-                        'count' => 10,
-                        'currentPage' => 1,
-                        'lastPage' => 2,
-                        'total' => 12,
-                    ],
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => '{
-                        builds(count: 10) {
-                            paginatorInfo {
-                                count
-                                currentPage
-                                lastPage
-                                total
-                            }
-                            data {
-                                id
-                                hash
-                                project {
-                                    id
-                                    name
-                                    repository
-                                }
-                                status
-                                commit
-                                completed_at
-                            }
-                        }
-                    }',
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testBuildQuery()
-    {
-        $build = factory(Build::class)->create();
-
-        $expected = [
-            'data' => [
-                'build' => [
-                    'id' => (string) $build->id,
-                    'hash' => $build->hash,
-                    'project' => [
-                        'id' => (string) $build->project->id,
-                        'name' => $build->project->name,
-                    ],
-                    'status' => $build->status,
-                    'commit' => $build->commit,
-                    'completed_at' => $build->completed_at,
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => "{
-                        build(hash: \"{$build->hash}\") {
-                            id
-                            hash
-                            project {
-                                id
-                                name
-                            }
-                            status
-                            commit
-                            completed_at
-                        }
-                    }",
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testProjectsQuery()
-    {
-        $projects = factory(Project::class, 12)
-            ->create()
-            ->take(10)
-            ->map(
-                function ($project) {
-                    return [
-                        'id' => (string) $project->id,
-                        'team' => [
-                            'id' => (string) $project->team->id,
-                            'name' => $project->team->name,
-                            'created_at' => $this->formatDateTime($project->team->created_at),
-                            'updated_at' => $this->formatDateTime($project->team->updated_at),
-                        ],
-                        'name' => $project->name,
-                        'repository' => $project->repository,
-                        'created_at' => $this->formatDateTime($project->created_at),
-                        'updated_at' => $this->formatDateTime($project->updated_at),
-                    ];
-                }
-            )
-            ->all();
-
-        $expected = [
-            'data' => [
-                'projects' => [
-                    'data' => $projects,
-                    'paginatorInfo' => [
-                        'count' => 10,
-                        'currentPage' => 1,
-                        'lastPage' => 2,
-                        'total' => 12,
-                    ],
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => '{
-                        projects(count: 10) {
-                            paginatorInfo {
-                                count
-                                currentPage
-                                lastPage
-                                total
-                            }
-                            data {
-                                id
-                                team {
-                                    id
-                                    name
-                                    created_at
-                                    updated_at
-                                }
-                                name
-                                repository
-                                created_at
-                                updated_at
-                            }
-                        }
-                    }',
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testProjectQuery()
-    {
-        $project = factory(Project::class)->create();
-
-        $expected = [
-            'data' => [
-                'project' => [
-                    'id' => (string) $project->id,
-                    'name' => $project->name,
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => "{
-                        project(id: {$project->id}) {
-                            id
-                            name
-                        }
-                    }",
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testTeamsQuery()
-    {
-        $teams = factory(Team::class, 12)
-            ->create()
-            ->take(10)
-            ->map(
-                function ($team) {
-                    return [
-                        'id' => (string) $team->id,
-                        'name' => $team->name,
-                    ];
-                }
-            )
-            ->all();
-
-        $expected = [
-            'data' => [
-                'teams' => [
-                    'data' => $teams,
-                    'paginatorInfo' => [
-                        'count' => 10,
-                        'currentPage' => 1,
-                        'lastPage' => 2,
-                        'total' => 12,
-                    ],
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => '{
-                        teams(count: 10) {
-                            paginatorInfo {
-                                count
-                                currentPage
-                                lastPage
-                                total
-                            }
-                            data {
-                                id
-                                name
-                            }
-                        }
-                    }',
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testTeamQuery()
-    {
-        $team = factory(Team::class)->create();
-
-        $expected = [
-            'data' => [
-                'team' => [
-                    'id' => (string) $team->id,
-                    'name' => $team->name,
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => "{
-                        team(id: {$team->id}) {
-                            id
-                            name
-                        }
-                    }",
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testUsersQuery()
-    {
-        factory(User::class, 11)->create();
-
-        $users = User::query()
-            ->take(10)
-            ->get()
-            ->map(
-                function ($user) {
-                    return [
-                        'id' => (string) $user->id,
-                        'name' => $user->name,
-                    ];
-                }
-            )
-            ->all();
-
-        $expected = [
-            'data' => [
-                'users' => [
-                    'data' => $users,
-                    'paginatorInfo' => [
-                        'count' => 10,
-                        'currentPage' => 1,
-                        'lastPage' => 2,
-                        'total' => 12,
-                    ],
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => '{
-                        users(count: 10) {
-                            paginatorInfo {
-                                count
-                                currentPage
-                                lastPage
-                                total
-                            }
-                            data {
-                                id
-                                name
-                            }
-                        }
-                    }',
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-
-    public function testUserQuery()
-    {
-        $user = factory(User::class)->create();
-
-        $expected = [
-            'data' => [
-                'user' => [
-                    'id' => (string) $user->id,
-                    'name' => $user->name,
-                    'email' => $user->email,
-                ],
-            ],
-        ];
-
-        $this
-            ->postJson(
-                '/graphql',
-                [
-                    'query' => "{
-                        user(id: {$user->id}) {
-                            id
-                            name
-                            email
-                        }
-                    }",
-                ]
-            )
-            ->assertStatus(200)
-            ->assertExactJson($expected);
-    }
-}
