+++ tests/Feature/TeamQueriesTest.php
@@ -11,6 +11,88 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class TeamQueriesTest extends PassportTestCase
 {
+    public function testTeamsQuery()
+    {
+        $teams = factory(Team::class, 12)
+            ->create()
+            ->take(10)
+            ->map(
+                function ($team) {
+                    return [
+                        'id' => (string) $team->id,
+                        'name' => $team->name,
+                    ];
+                }
+            )
+            ->all();
+
+        $expected = [
+            'data' => [
+                'teams' => [
+                    'data' => $teams,
+                    'paginatorInfo' => [
+                        'count' => 10,
+                        'currentPage' => 1,
+                        'lastPage' => 2,
+                        'total' => 12,
+                    ],
+                ],
+            ],
+        ];
+
+        $this
+            ->postJson(
+                '/graphql',
+                [
+                    'query' => '{
+                        teams(count: 10) {
+                            paginatorInfo {
+                                count
+                                currentPage
+                                lastPage
+                                total
+                            }
+                            data {
+                                id
+                                name
+                            }
+                        }
+                    }',
+                ]
+            )
+            ->assertStatus(200)
+            ->assertExactJson($expected);
+    }
+
+    public function testTeamQuery()
+    {
+        $team = factory(Team::class)->create();
+
+        $expected = [
+            'data' => [
+                'team' => [
+                    'id' => (string) $team->id,
+                    'name' => $team->name,
+                ],
+            ],
+        ];
+
+        $this
+            ->postJson(
+                '/graphql',
+                [
+                    'query' => "{
+                        team(id: {$team->id}) {
+                            id
+                            name
+                        }
+                    }",
+                ]
+            )
+            ->assertStatus(200)
+            ->assertExactJson($expected);
+    }
+
     public function testTeamProjectsQuery()
     {
         $otherProject = factory(Project::class);
